+++ Commons/src/Commons.DataContracts/Models/Force.cs
@@ -1,12 +1,29 @@
-﻿using System;
-
-namespace MudRunner.Commons.DataContracts.Models
+﻿namespace MudRunner.Commons.DataContracts.Models
 {
     /// <summary>
     /// It represents the force.
     /// </summary>
     public class Force
     {
+        /// <summary>
+        /// Basic constructor.
+        /// </summary>
+        public Force() { }
+
+        /// <summary>
+        /// Class constructor.
+        /// </summary>
+        /// <param name="x"></param>
+        /// <param name="y"></param>
+        /// <param name="z"></param>
+        public Force(double x, double y, double z)
+        {
+            this.AbsolutValue = Math.Sqrt(Math.Pow(x, 2) + Math.Pow(y, 2) + Math.Pow(z, 2));
+            this.X = x;
+            this.Y = y;
+            this.Z = z;
+        }
+
         /// <summary>
         /// the absolut value to force.
         /// </summary>
@@ -125,6 +142,24 @@ namespace MudRunner.Commons.DataContracts.Models
                 X = absolutValue * normalizedDirection.X,
                 Y = absolutValue * normalizedDirection.Y,
                 Z = absolutValue * normalizedDirection.Z
+            };
+        }
+
+        /// <summary>
+        /// This method creates the <see cref="Force"/> based on a string.
+        /// </summary>
+        /// <param name="force"></param>
+        /// <returns></returns>
+        public static Force Create(string force)
+        {
+            var vector3D = Vector3D.Create(force);
+
+            return new Force
+            {
+                AbsolutValue = vector3D.Length,
+                X = vector3D.X,
+                Y = vector3D.Y,
+                Z = vector3D.Z
             };
         }