+++ Suspension/src/Suspension.Core/Operations/RunAnalysis/Static/IRunStaticAnalysis.cs
@@ -41,12 +41,12 @@ namespace MudRunner.Suspension.Core.Operations.RunAnalysis
         Task<SingleComponentStaticAnalysisResult> GenerateSingleComponentResultAsync(SingleComponent<TProfile> component, bool shouldRoundResults, int decimals = 0);
 
         /// <summary>
-        /// Asynchronously, this method generates the analysis result to suspension A-arm.
+        /// Asynchronously, this method generates the analysis result to wishbone.
         /// </summary>
         /// <param name="component"></param>
         /// <param name="shouldRoundResults"></param>
         /// <param name="decimals"></param>
         /// <returns></returns>
-        Task<SuspensionAArmStaticAnalysisResult> GenerateSuspensionAArmResultAsync(SuspensionAArm<TProfile> component, bool shouldRoundResults, int decimals = 0);
+        Task<WishboneStaticAnalysisResult> GenerateWishboneResultAsync(Wishbone<TProfile> component, bool shouldRoundResults, int decimals = 0);
     }
 }