+++ Suspension/src/Suspension.Core/Operations/RunAnalysis/Static/RunStaticAnalysis.cs
@@ -81,12 +81,12 @@ namespace MudRunner.Suspension.Core.Operations.RunAnalysis
 
             tasks.Add(Task.Run(async () =>
             {
-                response.Data.SuspensionAArmLowerResult = await this.GenerateSuspensionAArmResultAsync(suspensionSystem.SuspensionAArmLower, request.ShouldRoundResults, request.NumberOfDecimalsToRound.GetValueOrDefault()).ConfigureAwait(false);
+                response.Data.LowerWishboneResult = await this.GenerateWishboneResultAsync(suspensionSystem.LowerWishbone, request.ShouldRoundResults, request.NumberOfDecimalsToRound.GetValueOrDefault()).ConfigureAwait(false);
             }));
 
             tasks.Add(Task.Run(async () =>
             {
-                response.Data.SuspensionAArmUpperResult = await this.GenerateSuspensionAArmResultAsync(suspensionSystem.SuspensionAArmUpper, request.ShouldRoundResults, request.NumberOfDecimalsToRound.GetValueOrDefault()).ConfigureAwait(false);
+                response.Data.UpperWishboneResult = await this.GenerateWishboneResultAsync(suspensionSystem.UpperWishbone, request.ShouldRoundResults, request.NumberOfDecimalsToRound.GetValueOrDefault()).ConfigureAwait(false);
             }));
 
             tasks.Add(Task.Run(async () =>
@@ -112,8 +112,8 @@ namespace MudRunner.Suspension.Core.Operations.RunAnalysis
                 AppliedForce = request.AppliedForce,
                 Origin = request.Origin,
                 ShockAbsorber = ShockAbsorberPoint.Create(request.ShockAbsorber),
-                SuspensionAArmLower = SuspensionAArmPoint.Create(request.SuspensionAArmLower),
-                SuspensionAArmUpper = SuspensionAArmPoint.Create(request.SuspensionAArmUpper),
+                LowerWishbone = WishbonePoint.Create(request.LowerWishbone),
+                UpperWishbone = WishbonePoint.Create(request.UpperWishbone),
                 TieRod = TieRodPoint.Create(request.TieRod)
             };
         }
@@ -140,7 +140,7 @@ namespace MudRunner.Suspension.Core.Operations.RunAnalysis
         /// <param name="shouldRoundResults"></param>
         /// <param name="decimals"></param>
         /// <returns></returns>
-        public async Task<SuspensionAArmStaticAnalysisResult> GenerateSuspensionAArmResultAsync(CoreModels.SuspensionAArm<TProfile> component, bool shouldRoundResults, int decimals = 0)
+        public async Task<WishboneStaticAnalysisResult> GenerateWishboneResultAsync(CoreModels.Wishbone<TProfile> component, bool shouldRoundResults, int decimals = 0)
         {
             if (component == null)
                 throw new ArgumentNullException(nameof(component), "The object suspension A-arm cannot be null to calculate the results.");
@@ -150,16 +150,16 @@ namespace MudRunner.Suspension.Core.Operations.RunAnalysis
                 FirstSegment = await GenerateSingleComponentResultAsync(new SingleComponent<TProfile>
                 {
                     AppliedForce = component.AppliedForce1,
-                    PivotPoint = component.PivotPoint1,
-                    FasteningPoint = component.KnucklePoint,
+                    PivotPoint = component.FrontPivot,
+                    FasteningPoint = component.OuterBallJoint,
                     Material = component.Material,
                     Profile = component.Profile,
                 }, shouldRoundResults, decimals).ConfigureAwait(false),
                 SecondSegment = await GenerateSingleComponentResultAsync(new SingleComponent<TProfile>
                 {
                     AppliedForce = component.AppliedForce2,
-                    PivotPoint = component.PivotPoint2,
-                    FasteningPoint = component.KnucklePoint,
+                    PivotPoint = component.RearPivot,
+                    FasteningPoint = component.OuterBallJoint,
                     Material = component.Material,
                     Profile = component.Profile,
                 }, shouldRoundResults, decimals).ConfigureAwait(false)