+++ Suspension/src/Suspension.DataContracts/CalculateReactions/CalculateReactionsRequest.cs
@@ -37,14 +37,14 @@ namespace MudRunner.Suspension.DataContracts.CalculateReactions
         public ShockAbsorberPoint ShockAbsorber { get; set; }
 
         /// <summary>
-        /// The suspension A-arm upper points.
+        /// The suspension upper wishbone points.
         /// </summary>
-        public SuspensionAArmPoint SuspensionAArmUpper { get; set; }
+        public WishbonePoint UpperWishbone { get; set; }
 
         /// <summary>
-        /// The suspension A-arm lower points.
+        /// The suspension lower wishbone points.
         /// </summary>
-        public SuspensionAArmPoint SuspensionAArmLower { get; set; }
+        public WishbonePoint LowerWishbone { get; set; }
 
         /// <summary>
         /// The tie rod points.