+++ Suspension/src/Suspension.DataContracts/CalculateSteeringKnuckleReactions/CalculateSteeringKnuckleReactionsResponse.cs
@@ -1,0 +1,9 @@
+﻿using MudRunner.Commons.DataContracts.Operation;
+
+namespace MudRunner.Suspension.DataContracts.CalculateSteeringKnuckleReactions
+{
+    /// <summary>
+    /// It represents the response content to CalculateSteeringKnuckleReactions operation.
+    /// </summary>
+    public class CalculateSteeringKnuckleReactionsResponse : OperationResponseBase<CalculateSteeringKnuckleReactionsResponseData> { }
+}