+++ Suspension/src/Suspension.DataContracts/RunAnalysis/Fatigue/RunFatigueAnalysisRequest.cs
@@ -93,12 +93,12 @@ namespace MudRunner.Suspension.DataContracts.RunAnalysis.Fatigue
         /// <summary>
         /// The suspension A-arm upper.
         /// </summary>
-        public SuspensionAArm<TProfile> SuspensionAArmUpper { get; set; }
+        public Wishbone<TProfile> UpperWishbone { get; set; }
 
         /// <summary>
         /// The suspension A-arm lower.
         /// </summary>
-        public SuspensionAArm<TProfile> SuspensionAArmLower { get; set; }
+        public Wishbone<TProfile> LowerWishbone { get; set; }
 
         /// <summary>
         /// The tie rod.