+++ Suspension/src/Suspension.DataContracts/RunAnalysis/Static/SuspensionAArmStaticAnalysisResult.cs
@@ -1,7 +1,0 @@
-﻿namespace MudRunner.Suspension.DataContracts.RunAnalysis.Static
-{
-    /// <summary>
-    /// It contains the essential static analysis results for a suspension A-arm.
-    /// </summary>
-    public class SuspensionAArmStaticAnalysisResult : SuspensionAArmAnalysisResult<SingleComponentStaticAnalysisResult> { }
-}