+++ Suspension/test/Suspension.UnitTest/Helper/CalculateReactionsHelper.cs
@@ -26,17 +26,17 @@ namespace MudRunner.Suspension.UnitTest.Helper
                     FasteningPoint = new Point3D { X = -0.005, Y = 0.645, Z = 0.180 },
                     PivotPoint = new Point3D { X = -0.005, Y = 0.485, Z = 0.430 }
                 },
-                SuspensionAArmLower = new SuspensionAArm
+                LowerWishbone = new Wishbone
                 {
-                    KnucklePoint = new Point3D { X = -0.012, Y = 0.685, Z = 0.150 },
-                    PivotPoint1 = new Point3D { X = -0.100, Y = 0.350, Z = 0.130 },
-                    PivotPoint2 = new Point3D { X = 0.250, Y = 0.350, Z = 0.150 }
+                    OuterBallJoint = new Point3D { X = -0.012, Y = 0.685, Z = 0.150 },
+                    FrontPivot = new Point3D { X = -0.100, Y = 0.350, Z = 0.130 },
+                    RearPivot = new Point3D { X = 0.250, Y = 0.350, Z = 0.150 }
                 },
-                SuspensionAArmUpper = new SuspensionAArm
+                UpperWishbone = new Wishbone
                 {
-                    KnucklePoint = new Point3D { X = 0.012, Y = 0.660, Z = 0.410 },
-                    PivotPoint1 = new Point3D { X = -0.080, Y = 0.450, Z = 0.362 },
-                    PivotPoint2 = new Point3D { X = 0.200, Y = 0.450, Z = 0.362 }
+                    OuterBallJoint = new Point3D { X = 0.012, Y = 0.660, Z = 0.410 },
+                    FrontPivot = new Point3D { X = -0.080, Y = 0.450, Z = 0.362 },
+                    RearPivot = new Point3D { X = 0.200, Y = 0.450, Z = 0.362 }
                 },
                 TieRod = new TieRod
                 {
@@ -59,17 +59,17 @@ namespace MudRunner.Suspension.UnitTest.Helper
                     FasteningPoint = "-0.005,0.645,0.180",
                     PivotPoint = "-0.005,0.485,0.430"
                 },
-                SuspensionAArmLower = new SuspensionAArmPoint
+                LowerWishbone = new WishbonePoint
                 {
-                    KnucklePoint = "-0.012,0.685,0.150",
-                    PivotPoint1 = "-0.100,0.350,0.130",
-                    PivotPoint2 = "0.250,0.350,0.150"
+                    OuterBallJoint = "-0.012,0.685,0.150",
+                    FrontPivot = "-0.100,0.350,0.130",
+                    RearPivot = "0.250,0.350,0.150"
                 },
-                SuspensionAArmUpper = new SuspensionAArmPoint
+                UpperWishbone = new WishbonePoint
                 {
-                    KnucklePoint = "0.012,0.660,0.410",
-                    PivotPoint1 = "-0.080,0.450,0.362",
-                    PivotPoint2 = "0.200,0.450,0.362"
+                    OuterBallJoint = "0.012,0.660,0.410",
+                    FrontPivot = "-0.080,0.450,0.362",
+                    RearPivot = "0.200,0.450,0.362"
                 },
                 TieRod = new TieRodPoint
                 {
@@ -146,19 +146,19 @@ namespace MudRunner.Suspension.UnitTest.Helper
         {
             return new CalculateReactionsResponseData
             {
-                AArmLowerReaction1 = new Force
+                LowerWishboneReaction1 = new Force
                 {
                     AbsolutValue = -706.844136886457
                 },
-                AArmLowerReaction2 = new Force
+                LowerWishboneReaction2 = new Force
                 {
                     AbsolutValue = 2318.54871728814
                 },
-                AArmUpperReaction1 = new Force
+                UpperWishboneReaction1 = new Force
                 {
                     AbsolutValue = 410.390832183452
                 },
-                AArmUpperReaction2 = new Force
+                UpperWishboneReaction2 = new Force
                 {
                     AbsolutValue = -693.435739188224
                 },